Revert "Refactor Geocoding methods"
